+++ push/sender/src/main/java/org/jboss/aerogear/unifiedpush/message/SenderServiceImpl.java
@@ -19,12 +19,8 @@ package org.jboss.aerogear.unifiedpush.message;
 import org.jboss.aerogear.unifiedpush.api.PushApplication;
 import org.jboss.aerogear.unifiedpush.api.PushMessageInformation;
 import org.jboss.aerogear.unifiedpush.api.Variant;
-import org.jboss.aerogear.unifiedpush.api.VariantMetricInformation;
 import org.jboss.aerogear.unifiedpush.api.VariantType;
 import org.jboss.aerogear.unifiedpush.utils.AeroGearLogger;
-import org.jboss.aerogear.unifiedpush.message.sender.NotificationSenderCallback;
-import org.jboss.aerogear.unifiedpush.message.sender.PushNotificationSender;
-import org.jboss.aerogear.unifiedpush.service.ClientInstallationService;
 import org.jboss.aerogear.unifiedpush.service.GenericVariantService;
 import org.jboss.aerogear.unifiedpush.service.metrics.PushMessageMetricsService;
 
@@ -33,8 +29,6 @@ import javax.ejb.Asynchronous;
 import javax.ejb.Stateless;
 import javax.ejb.TransactionAttribute;
 import javax.ejb.TransactionAttributeType;
-import javax.enterprise.inject.Any;
-import javax.enterprise.inject.Instance;
 import javax.inject.Inject;
 import javax.jms.Connection;
 import javax.jms.ConnectionFactory;
@@ -56,12 +50,6 @@ public class SenderServiceImpl implements SenderService {
 
     private final AeroGearLogger logger = AeroGearLogger.getInstance(SenderServiceImpl.class);
 
-    @Inject
-    @Any
-    private Instance<PushNotificationSender> senders;
-
-    @Inject
-    private ClientInstallationService clientInstallationService;
     @Inject
     private GenericVariantService genericVariantService;
     @Inject
@@ -110,67 +98,28 @@ public class SenderServiceImpl implements SenderService {
             variants.addAll(pushApplication.getVariants());
         }
 
-        // we split the variants per type since each type has its own
-        Connection connection;
+        Connection connection = null;
         try {
             connection = connectionFactory.createConnection();
             Session session = connection.createSession(false, Session.AUTO_ACKNOWLEDGE);
             MessageProducer messageProducer = session.createProducer(variantTypeQueue);
             connection.start();
 
+            // we split the variants per type since each type may have its own configuration (e.g. batch size)
             for (final Entry<VariantType, List<Variant>> variant : variants.entrySet()) {
-    //            final List<String> tokenPerVariant = clientInstallationService.findAllDeviceTokenForVariantIDByCriteria(variant.getVariantID(), categories, aliases, deviceTypes);
-    //            senders.select(new SenderTypeLiteral(variant.getClass())).get()
-    //                    .sendPushMessage(variant, tokenPerVariant, message, new SenderServiceCallback(variant, tokenPerVariant.size(), pushMessageInformation));
                 ObjectMessage messageWithVariants = session.createObjectMessage(new MessageForVariants(message, variant.getValue()));
                 messageProducer.send(messageWithVariants);
             }
         } catch (JMSException e) {
             throw new MessageDeliveryException("Failed to queue push message for further processing", e);
-        }
-    }
-
-    /**
-     * Helpers to update the given {@link PushMessageInformation} with a {@link VariantMetricInformation} object
-     */
-    private void updateStatusOfPushMessageInformation(final PushMessageInformation pushMessageInformation, final String variantID, final int receives, final Boolean deliveryStatus) {
-        this.updateStatusOfPushMessageInformation(pushMessageInformation, variantID, receives, deliveryStatus, null);
-    }
-
-    private void updateStatusOfPushMessageInformation(final PushMessageInformation pushMessageInformation, final String variantID, final int receives, final Boolean deliveryStatus, final String reason) {
-        final VariantMetricInformation variantMetricInformation = new VariantMetricInformation();
-        variantMetricInformation.setVariantID(variantID);
-        variantMetricInformation.setReceivers(receives);
-        variantMetricInformation.setDeliveryStatus(deliveryStatus);
-        variantMetricInformation.setReason(reason);
-        pushMessageInformation.addVariantInformations(variantMetricInformation);
-        pushMessageInformation.setTotalReceivers(pushMessageInformation.getTotalReceivers() + receives);
-
-        // store it!
-        metricsService.updatePushMessageInformation(pushMessageInformation);
-    }
-
-    private class SenderServiceCallback implements NotificationSenderCallback {
-        private final Variant variant;
-        private final int tokenSize;
-        private final PushMessageInformation pushMessageInformation;
-
-        public SenderServiceCallback(Variant variant, int tokenSize, PushMessageInformation pushMessageInformation) {
-            this.variant = variant;
-            this.tokenSize = tokenSize;
-            this.pushMessageInformation = pushMessageInformation;
-        }
-
-        @Override
-        public void onSuccess() {
-            logger.fine(String.format("Sent '%s' message to '%d' devices", variant.getType().getTypeName(), tokenSize));
-            updateStatusOfPushMessageInformation(pushMessageInformation, variant.getVariantID(), tokenSize, Boolean.TRUE);
-        }
-
-        @Override
-        public void onError(final String reason) {
-            logger.warning(String.format("Error on '%s' delivery", variant.getType().getTypeName()));
-            updateStatusOfPushMessageInformation(pushMessageInformation, variant.getVariantID(), tokenSize, Boolean.FALSE, reason);
+        } finally {
+            if (connection != null) {
+                try {
+                    connection.close();
+                } catch (JMSException e) {
+                    e.printStackTrace();
+                }
+            }
         }
     }
